Move docker/tarfile.Source to docker/internal/tarfile.Source
... so that we can add more functionality to it without making it
public API.

Keep the original docker/tarfile.Source as a compatibility shim.

Similarly, move docker/tarfile/src_test.go as well, to keep it
close to the implementation being tested.

Should not change behavior.
