+++ docker/tarfile/types.go
@@ -4,12 +4,5 @@ import (
 	internal "github.com/containers/image/v5/docker/internal/tarfile"
 )
 
-// Various data structures.
-
-// Based on github.com/docker/docker/image/tarexport/tarexport.go
-const (
-	manifestFileName = "manifest.json"
-)
-
 // ManifestItem is an element of the array stored in the top-level manifest.json file.
 type ManifestItem = internal.ManifestItem // All public members from the internal package remain accessible.
